requirements.txt
tweepy
requests 
python-dotenv